+++ src/Admin/State/AdminMarketingCartRuleCouponGenerateProcessor.php
@@ -178,7 +178,7 @@ class AdminMarketingCartRuleCouponGenerateProcessor implements ProcessorInterfac
     {
         return new JsonResponse(array_filter([
             'id'         => $dto->id,
-            'cartRuleId' => $dto->cartRuleId,
+            'cartRuleId' => $dto->cart_rule_id,
             'generated'  => $dto->generated,
             'coupons'    => $dto->coupons,
             'success'    => $dto->success,